Rename folders, change numeric types

+++ WPFClient/Database/Employee.cs
@@ -2,14 +2,14 @@
 {
     public class Employee
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string FatherName { get; set; }
         public string Position { get; set; }
-        public double Salary { get; set; }
+        public decimal Salary { get; set; }
 
-        public int DepartmentId { get; set; }
+        public ulong DepartmentId { get; set; }
         public virtual Department Department { get; set; }
     }
 }
